Variable scope supported
- Add scope for variable. The rules are follow:
    - Each block or function is a scope.
   - Variable in the child scope can has the same name as a variable in the other scope. In that case, when an operation reference a variable that has the same name in both scope, the nearest scope variable will be taken, as the variable searching algorithm is depth recursive.
    - TODO: Record (Class) and global variable scope.

- Adjust index operation to extend index range to a 2-bits integer

+++ src/ir_compiler.cpp
@@ -82,7 +82,7 @@ namespace snack::ir::backend {
         case opcode::strlc:
         case opcode::vri:
         case opcode::vrs: {
-            uint8_t idx = static_cast<uint8_t>(value);
+            uint16_t idx = static_cast<uint16_t>(value);
 
             ir_bin.write(reinterpret_cast<const char *>(&idx), 1);
             funcs.back().crr_pc += 1;
